+++ features/step_definitions/version.ts
@@ -1,5 +1,5 @@
 import * as assert from "assert";
-import { Then } from "../ultils/world";
+import { Then } from "../supports/world";
 
 Then(
   "I should get the version formatted in {string}",
